Fix exclude/include.
ConventionMapping only works when a getter is explicitly called.

SourceTask.getSource() does not call getExcludes() or getIncludes(),
instead it uses an internal patternSet. This is a workaround to update
the SourceTask patternSet using the values provided by the convention
mapping before calling getSource().

There is probably a better solution, but it probably means using
something other than ConventionMapping for the SourceTask's
include/exclude.

Also, exclude now applies to the instrumentation tasks as well.